move groups into feature folder
