+++ papermerge/core/db/__init__.py
@@ -20,7 +20,6 @@ from .doc_ver import get_doc_ver, get_last_doc_ver
 from .engine import get_engine
 from .exceptions import UserNotFound
 from .folders import get_folder
-from .groups import create_group, delete_group, get_group, get_groups, update_group
 from .nodes import get_nodes, get_paginated_nodes
 from .pages import get_doc_ver_pages, get_first_page, get_page
 from .perms import get_perms, sync_perms
@@ -56,11 +55,6 @@ __all__ = [
     "Engine",
     "Session",
     "UserNotFound",
-    "get_group",
-    "get_groups",
-    "create_group",
-    "update_group",
-    "delete_group",
     "get_nodes",
     "get_custom_fields",
     "create_custom_field",
