+++ papermerge/core/db/users.py
@@ -9,7 +9,8 @@ from sqlalchemy.exc import NoResultFound
 from papermerge.core import constants, schemas
 from papermerge.core.auth import scopes
 from papermerge.core.db.engine import Session
-from papermerge.core.db.models import Folder, Group, Permission, User
+from papermerge.core.db.models import Folder, User
+from papermerge.core.features.groups.db.orm import Group, Permission
 from papermerge.core.utils.misc import is_valid_uuid
 
 from .exceptions import UserNotFound
